Fix every -Wsign-conversion site and turn the warning on for good (#172)
The previous commit left this deferred behind an off-by-default option.
That was the wrong call: #172 says there shouldn't be any warnings, and
~3200 of them were sitting in the one diagnostic the gate compiler
happened not to enable.

-Wsign-conversion is now named explicitly in project_warnings, so GCC and
Clang audit the same thing, and all 1709 unique sites across 172 files are
fixed: 3242 GCC diagnostics and 3231 Clang diagnostics down to zero.

Every edit makes a conversion the compiler was already performing explicit,
around the whole converted expression so the arithmetic still happens in
its original type. No declaration, signature, member type or loop-counter
type changed, and no value changed. Every one of the 172 files is exactly
line-count preserving (git numstat is N/N for all of them), which is what
keeps the 217 line-anchored mutation-canary pins valid.

Verified warning-free under GCC 15.2, GCC 13.4 (CI's compiler) and
Clang 21.1, and building clean under ci-test, ci-asan and ci-fuzz with
-Werror on. Behavior is unchanged where it is measurable to the byte:
og_test_parity 256/256 with tests/parity/golden/ untouched, ctest 58/58,
mutation pins 217 valid.

The sweep surfaced nine latent defects where the converted value can
genuinely be negative or out of range. All nine are cast to preserve
today's behavior and listed in the PR — fixing them changes behavior and
belongs in its own change with its own tests, not in a warning cleanup.

+++ src/core/combat_math.cpp
@@ -155,7 +155,7 @@ short compute_xp_from_action(ExpAction action, std::int32_t attacker_level, std:
         return compute_xp_from_kill(level_diff);
     case ExpAction::Heal: {
         const std::int32_t denom = (attacker_level > 0) ? attacker_level : 1;
-        return static_cast<short>(rng.next(static_cast<std::uint32_t>(20 * value)) / denom);
+        return static_cast<short>(rng.next(static_cast<std::uint32_t>(20 * value)) / static_cast<std::uint32_t>(denom));
     }
     case ExpAction::TurnUndead:
         return static_cast<short>(value * 3);
