+++ src/gameplay/walker_specials.cpp
@@ -335,7 +335,7 @@ std::int32_t walker::turn_undead(std::int32_t range, [[maybe_unused]] std::int32
 		const auto* target_fd = w ? get_family_descriptor(w->family()) : nullptr;
 		if (w && target_fd && target_fd->is_undead)
 		{
-			if (current_game->world->rng_.next(range*40) > current_game->world->rng_.next(w->stats()->level()*10) )
+			if (current_game->world->rng_.next(static_cast<std::uint32_t>(range*40)) > current_game->world->rng_.next(static_cast<std::uint32_t>(w->stats()->level()*10)) )
 			{
 				GameWorld* const world = current_game->world;
 				const bool scripted_mode =
